+++ kitchen-sink/vue/src/pages/login-screen-page.vue
@@ -2,14 +2,20 @@
   <f7-page no-toolbar no-navbar no-swipeback login-screen>
     <f7-login-screen-title>Framework7</f7-login-screen-title>
     <f7-list form>
-      <f7-list-item>
-        <f7-label>Username</f7-label>
-        <f7-input type="text" placeholder="Your username" @input="username = $event.target.value"></f7-input>
-      </f7-list-item>
-      <f7-list-item>
-        <f7-label>Password</f7-label>
-        <f7-input type="password" placeholder="Your password" @input="password = $event.target.value"></f7-input>
-      </f7-list-item>
+      <f7-list-input
+        label="Username"
+        type="text"
+        placeholder="Your username"
+        :value="username"
+        @input="username = $event.target.value"
+      ></f7-list-input>
+      <f7-list-input
+        label="Password"
+        type="password"
+        placeholder="Your password"
+        :value="password"
+        @input="password = $event.target.value"
+      ></f7-list-input>
     </f7-list>
     <f7-list>
       <f7-list-button @click="signIn">Sign In</f7-list-button>
@@ -18,7 +24,7 @@
   </f7-page>
 </template>
 <script>
-  import { f7Page, f7LoginScreenTitle, f7List, f7ListItem, f7Label, f7Input, f7ListButton, f7BlockFooter } from 'framework7-vue';
+  import { f7Page, f7LoginScreenTitle, f7List, f7ListItem, f7ListButton, f7BlockFooter, f7ListInput } from 'framework7-vue';
 
   export default {
     components: {
@@ -26,10 +32,9 @@
       f7LoginScreenTitle,
       f7List,
       f7ListItem,
-      f7Label,
-      f7Input,
       f7ListButton,
       f7BlockFooter,
+      f7ListInput,
     },
     data() {
       return {