+++ packages/vue/components/messagebar.js
@@ -24,7 +24,10 @@ export default {
       default: 0
     },
     maxHeight: Number,
-    resizePage: Boolean,
+    resizePage: {
+      type: Boolean,
+      default: true
+    },
     sendLink: String,
     value: [String, Number, Array],
     disabled: Boolean,