+++ src/phenome/components/list.jsx
@@ -86,10 +86,13 @@ export default {
           tag === 'li'
           || tag === 'F7ListItem'
           || tag === 'F7ListButton'
+          || tag === 'F7ListInput'
           || tag.indexOf('list-item') >= 0
           || tag.indexOf('list-button') >= 0
+          || tag.indexOf('list-input') >= 0
           || tag.indexOf('f7-list-item') >= 0
           || tag.indexOf('f7-list-button') >= 0
+          || tag.indexOf('f7-list-input') >= 0
         ))
       ) {
         if (wasUlChild) rootChildrenAfterList.push(child);