+++ src/phenome/components/messagebar.jsx
@@ -36,7 +36,10 @@ export default {
       default: 0,
     },
     maxHeight: Number,
-    resizePage: Boolean,
+    resizePage: {
+      type: Boolean,
+      default: true,
+    },
     sendLink: String,
     value: [String, Number, Array],
     disabled: Boolean,